requirements.txt
# Telegram Bot Framework
python-telegram-bot>=20.0

# Google Services
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# AI and NLP
ollama>=0.1.0
dateparser

# Web Scraping
requests
beautifulsoup4
lxml

# Utilities
python-dotenv
APScheduler
pytz

# Type hints support
typing-extensions
